Flytt typer for aktivitet til egen fil
