+++ src/frontend/Sider/Behandling/Inngangsvilkår/typer.ts
@@ -1,3 +1,4 @@
+import { Aktivitet, AktivitetType } from './typer/aktivitet';
 import { Målgruppe, MålgruppeType } from './typer/målgruppe';
 import { Periode } from '../../../utils/periode';
 import { Vilkårsresultat } from '../vilkår';
@@ -29,23 +30,6 @@ export enum SvarJaNei {
 export interface Vurdering {
     svar: SvarJaNei;
     resultat: Vilkårsresultat;
-}
-
-export interface Aktivitet extends VilkårPeriode {
-    id: string;
-    type: AktivitetType;
-    detaljer: DelvilkårAktivitet;
-}
-
-export enum AktivitetType {
-    TILTAK = 'TILTAK',
-    UTDANNING = 'UTDANNING',
-    REEL_ARBEIDSSØKER = 'REEL_ARBEIDSSØKER',
-}
-
-interface DelvilkårAktivitet {
-    lønnet: Vurdering;
-    mottarSykepenger: Vurdering;
 }
 
 export interface Stønadsperiode extends Periode {
